Stories are now generated in PDF file
