+++ README.md
@@ -1,4 +1,21 @@
 AgileStoryPrint
 ===============
+Because we think visual management is essential in you daily team management, 
+we develop a tool to allow you printing Jira Story to beautiful formated Story
+cards.
 
-Scrum user stories generator, helper for visual management
+Demo on: [www.jira2scrumboard.com](http://www.jira2scrumboard.com/)
+
+Minimum Requirements
+===============
+To run this project you need:
+* PHP 5.6
+* [Composer](https://getcomposer.org/)
+* [WKHTMLToPDF](http://wkhtmltopdf.org/downloads.html)
+
+The project is based on Symfony 2.6
+
+Contribution
+===============
+Feel free to contribute to this project and improve it.  
+Any new translation is welcomed.
