+++ app/config/config.yml
@@ -47,27 +47,6 @@ assetic:
             jar: "%kernel.root_dir%/Resources/java/yuicompressor-2.4.8.jar"
     #assets:
 
-# Doctrine Configuration
-# doctrine:
-    # dbal:
-        # driver:   "%database_driver%"
-        # host:     "%database_host%"
-        # port:     "%database_port%"
-        # dbname:   "%database_name%"
-        # user:     "%database_user%"
-        # password: "%database_password%"
-        # charset:  UTF8
-        # if using pdo_sqlite as your database driver:
-        #   1. add the path in parameters.yml
-        #     e.g. database_path: "%kernel.root_dir%/data/data.db3"
-        #   2. Uncomment database_path in parameters.yml.dist
-        #   3. Uncomment next line:
-        #     path:     "%database_path%"
-
-    # orm:
-        # auto_generate_proxy_classes: "%kernel.debug%"
-        # auto_mapping: true
-
 # Swiftmailer Configuration
 swiftmailer:
     transport: "%mailer_transport%"
@@ -75,3 +54,14 @@ swiftmailer:
     username:  "%mailer_user%"
     password:  "%mailer_password%"
     spool:     { type: memory }
+
+# HTML to PDF configuration
+knp_snappy:
+    pdf:
+        enabled:    true
+        binary:     %wkhtmltopdf_path%
+        options:    {'page-size':'A4', 'orientation':'landscape', 'margin-bottom':'5', 'margin-left':'10', 'margin-right':'10', 'margin-top':'5'}
+    image:
+        enabled:    false
+        binary:     %wkhtmltoimage_path%
+        options:    []
