+++ app/config/parameters.yml.dist
@@ -1,22 +1,23 @@
 # This file is a "template" of what your parameters.yml file should look like
 parameters:
-    # database_driver:      pdo_mysql
-    # database_host:        127.0.0.1
-    # database_port:        ~
-    # database_name:        symfony
-    # database_user:        root
-    # database_password:    ~
-    # You should uncomment this if you want use pdo_sqlite
-    # database_path: "%kernel.root_dir%/data.db3"
-
+    # Mailer setup
     mailer_transport:       smtp
     mailer_host:            127.0.0.1
     mailer_user:            ~
     mailer_password:        ~
+    # email adress where to send contact messages
     mailer_admin_address:   ~
+    # email address on email sent (ex: noreply@jira2scrumboard.com)
     mailer_noreply_address: ~
+    # email name on email sent (ex: Jira2ScrumBoard)
     mailer_noreply_name:    ~
 
+
+    # WKHTMLtoPDF & WHTMLtoImage (Did you install it on your server?)
+    wkhtmltopdf_path:       /usr/local/bin/wkhtmltopdf
+    wkhtmltoimage_path:       /usr/local/bin/wkhtmltoimage
+
+    # Default locale
     locale:                 en
 
     # A secret key that's used to generate certain security-related tokens
