+++ composer.json
@@ -21,7 +21,9 @@
         "incenteev/composer-parameter-handler": "~2.0",
         "symfony/dom-crawler": "3.0.*@dev",
         "besimple/i18n-routing-bundle": "dev-master",
-        "liuggio/excelbundle": "~2.0"
+        "liuggio/excelbundle": "~2.0",
+        "knplabs/knp-snappy-bundle": "dev-master",
+        "knplabs/knp-snappy": "0.3.*"
 
     },
     "require-dev": {
