+++ src/AgileStoryPrint/JiraBundle/Resources/config/routing_i18n.yml
@@ -4,15 +4,15 @@ _index:
 
 _getStories:
     locales:     { en: "/print-your-index-cards.html", fr: "/fr/imprimez-vos-index-cards.html" }
-    defaults: { _controller: AgileStoryPrintJiraBundle:Default:getStories }
+    defaults: { _controller: AgileStoryPrintJiraBundle:Stories:getStories }
 
 _emptyStory:
     locales:     { en: "/no-stories-found.html", fr: "/fr/pas-de-story-trouve.html" }
-    defaults: { _controller: AgileStoryPrintJiraBundle:Default:emptyStory }
+    defaults: { _controller: AgileStoryPrintJiraBundle:Stories:emptyStory }
 
 _wrongFile:
     locales:     { en: "/wrong-file.html", fr: "/fr/mauvais-fichier.html" }
-    defaults: { _controller: AgileStoryPrintJiraBundle:Default:wrongFile }
+    defaults: { _controller: AgileStoryPrintJiraBundle:Stories:wrongFile }
 
 _contact:
     locales:     { en: "/contact.html", fr: "/fr/contact.html" }
