--- src/AgileStoryPrint/JiraBundle/Resources/views/Pages/complete.html.twig
@@ -1,26 +1,0 @@
-{% extends 'AgileStoryPrintJiraBundle::base.html.twig' %}
-
-{% block content %}
-	{% include 'AgileStoryPrintJiraBundle::navigation.html.twig' with {'links': 'absolute'} %}
-
-	<!-- Header -->
-	<header>
-	    <div class="container">
-	        <div class="intro-text message">
-	            <div class="intro-lead-in">{{ 'complete.title'|trans }}</div>
-	            <div class="print-it">
-		            <a href="javascript:window.print();">
-			            <span class="fa-stack fa-8x">
-		                    <i class="fa fa-circle fa-stack-2x text-primary"></i>
-		                    <i class="fa fa-print fa-stack-1x fa-inverse"></i>
-		                </span>
-	            	</a>
-	            </div>
-	            <div class="intro-heading">{{ 'complete.text'|trans|raw }}</div>
-	        </div>
-	    </div>
-	</header>
-	<!-- /Header -->
-
-	{% include 'AgileStoryPrintJiraBundle:StoryCard:card.html.twig' %}
-{% endblock %}
